EOS-13330 net/klnet_core: add a list of hardcoded lnids for LDR R1
Currently we have a list of LNIDs, which is cached during m0tr.ko module
insertion. During a failover LNIDs could be added in runtime, which will
lift a requirement to re-insert m0tr.ko at that time. This may speed up
the failover by having Motr processes, that are running already on the
system, to continue to run without restart.

Current implementation just adds a hardcoded list of all possible LNIDs
when m0tr.ko is inserted, so there will be no need to implement cache
invalidation when a failover happens. This is a solution that could be
implemented very fast, and it will unblock further testing.

Signed-off-by: Maksym Medvied <max.medved@seagate.com>